Script Install DB Container

+++ mysql/startup.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
-f [ ! -f /var/lib/mysql/rancher]; then
+if [ ! -f /var/lib/mysql/rancherdb ]; then
 
-        mysql_install_db
+	mysql_install_db
 
-        /usr/bin/mysqld_safe &
-        sleep 10s
+	/usr/bin/mysqld_safe &
+	sleep 10s
+        
         echo "CREATE DATABASE rancher COLLATE = 'utf8_general_ci' CHARACTER SET = 'utf8'; GRANT ALL PRIVILEGES ON *.* TO 'rancher'@'localhost' IDENTIFIED BY 'rancher' WITH GRANT OPTION;" | mysql
         killall mysqld
         sleep 10s
